refactor: global/success 폴더 구조 적용 (#29)
