+++ src/main/java/eightbit/moyeohaeng/domain/auth/controller/AuthController.java
@@ -10,7 +10,7 @@ import eightbit.moyeohaeng.domain.auth.dto.request.SignUpRequest;
 import eightbit.moyeohaeng.domain.auth.dto.response.TokenResponse;
 import eightbit.moyeohaeng.domain.auth.exception.AuthSuccessCode;
 import eightbit.moyeohaeng.domain.auth.service.AuthService;
-import eightbit.moyeohaeng.global.success.common.SuccessResponse;
+import eightbit.moyeohaeng.global.success.SuccessResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
